// src/components/Plans/simpleFeaturedPlans/Interfaces/HeadSubHeadButtonStyles.tsx
import React from "react";

export interface HeadSubHeadButtonStyle {
  contentStyle: React.CSSProperties;
  mainContentStyle: React.CSSProperties;
  heading: string;
  subHeading: string;
  subHeadingStyle: React.CSSProperties;
  buttons: boolean;
  buttonsStyleDiv: React.CSSProperties;
  button1: {};
  button2: {};
}